# Add check for shop/office/craft too close to a way