Implements #1819

Adds a check for `shop`, `craft` or `office` entries as nodes that are too close to a way. (Buildings tagged with shop/office/craft overlapping with highways are already caught).

Implementation as close as possible to tree & power object too close to way, with a primary difference that I include the buffer zone in the `&&` check (otherwise a node close to an end point of a curved way could get lost)

---

Unrelated, but it may be useful to rename table `highway` (`sql00`) to `highway_railway`. I confused myself a couple of times with the generic table `highways`, but table `highway` includes some railways too (but is missing [`light_rail`](https://wiki.openstreetmap.org/wiki/Tag:railway%3Dlight_rail) I think). Also I suspect line 482 should be `self.create_view_touched("water", "W")` instead of `..., "N")`. And `elevator` is whitelisted twice in sql00 and sql10. But this can all be a future PR.

---

Some results in the details link below:
<details>

**Slovenia**:
```
<error class="12">
<node id="3438934568" lat="46.566022" lon="15.943065" version="2" user="">
<tag k="name" v="Zavarovalnica Maribor" />
<tag k="office" v="insurance" />
<tag k="addr:city" v="Cerkvenjak" />
<tag k="addr:street" v="Cerkvenjak" />
<tag k="addr:postcode" v="2236" />
</node>
<way id="336870964" version="6" user="">
<tag k="ref" v="734" />
<tag k="name" v="Cerkvenjak" />
<tag k="highway" v="tertiary" />
<tag k="surface" v="asphalt" />
<tag k="maxspeed" v="30" />
</way>
<location lat="46.566022" lon="15.943065" />
</error>
<error class="12">
<node id="5860680485" lat="46.2329506" lon="15.6376553" version="1" user="">
<tag k="name" v="Legro" />
<tag k="shop" v="shoes" />
</node>
<way id="847741302" version="1" user="">
<tag k="name" v="Strma cesta" />
<tag k="highway" v="residential" />
<tag k="surface" v="asphalt" />
</way>
<location lat="46.2329506" lon="15.6376553" />
</error>
```

**Japan_Chubu**:
```
<error class="12">
<node id="828448475" lat="34.9747325" lon="138.3834197" version="2" user="higa4">
<tag k="name" v="きものの成田屋 (Naritaya)" />
<tag k="shop" v="clothes" />
</node>
<way id="69080689" version="8" user="pon_de_lions">
<tag k="oneway" v="yes" />
<tag k="highway" v="unclassified" />
</way>
<location lat="34.9747325" lon="138.3834197" />
</error>
<error class="12">
<node id="1243957814" lat="38.133866" lon="139.426425" version="3" user="niityhksn">
<tag k="shop" v="金屋" />
<tag k="highway" v="traffic_signals" />
<tag k="traffic_signals" v="signal" />
</node>
<way id="108412122" version="6" user="niityhksn">
<tag k="ref" v="182" />
<tag k="name" v="坂町停車場金屋線" />
<tag k="highway" v="secondary" />
</way>
<location lat="38.133866" lon="139.426425" />
</error>
<error class="12">
<node id="1668368572" lat="36.332737" lon="137.2983148" version="1" user="nobori">
<tag k="shop" v="clothes" />
</node>
<way id="114342951" version="2" user="tomo_klotho">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="tertiary" />
</way>
<location lat="36.332737" lon="137.2983148" />
</error>
<error class="12">
<node id="1668368578" lat="36.3326959" lon="137.2986474" version="1" user="nobori">
<tag k="shop" v="toys" />
</node>
<way id="114344778" version="2" user="芋けんぴ">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="tertiary" />
</way>
<location lat="36.3326959" lon="137.2986474" />
</error>
<error class="12">
<node id="1668368583" lat="36.33267" lon="137.2989024" version="1" user="nobori">
<tag k="shop" v="music" />
</node>
<way id="114344778" version="2" user="芋けんぴ">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="tertiary" />
</way>
<location lat="36.33267" lon="137.2989024" />
</error>
<error class="12">
<node id="1668368584" lat="36.3327607" lon="137.2981594" version="1" user="nobori">
<tag k="shop" v="hairdresser" />
</node>
<way id="114337000" version="3" user="芋けんぴ">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="tertiary" />
</way>
<location lat="36.3327607" lon="137.2981594" />
</error>
<error class="12">
<node id="1668368588" lat="36.3326419" lon="137.2991571" version="1" user="nobori">
<tag k="shop" v="gift" />
</node>
<way id="114336856" version="1" user="Tom_G3X">
<tag k="ref" v="477" />
<tag k="name" v="長倉神岡線" />
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="secondary" />
<tag k="yh:TYPE" v="一般都道府県道" />
<tag k="yh:WIDTH" v="5.5m〜13.0m" />
<tag k="yh:LINE_NUM" v="477" />
<tag k="yh:LINE_NAME" v="長倉神岡線" />
<tag k="yh:STRUCTURE" v="地上" />
<tag k="yh:TOTYUMONO" v="供用中" />
<tag k="yh:WIDTH_RANK" v="2" />
</way>
<location lat="36.3326419" lon="137.2991571" />
</error>
<error class="12">
<node id="3287101801" lat="37.6485805" lon="138.9391245" version="1" user="muramototomoya">
<tag k="name" v="燕三条WING" />
<tag k="shop" v="hardware" />
</node>
<way id="24356184" version="12" user="flierfy">
<tag k="name" v="上越新幹線" />
<tag k="gauge" v="1435" />
<tag k="usage" v="main" />
<tag k="source" v="KSJ2" />
<tag k="name:de" v="Jōetsu-Shinkansen" />
<tag k="name:en" v="Joetsu Shinkansen" />
<tag k="name:fr" v="Ligne Shinkansen Jōetsu" />
<tag k="name:hu" v="Dzsóecu Sinkanszen" />
<tag k="name:ja" v="上越新幹線" />
<tag k="name:ru" v="Дзёэцу-синкансэн" />
<tag k="railway" v="rail" />
<tag k="voltage" v="25000" />
<tag k="KSJ2:LIN" v="上越新幹線" />
<tag k="maxspeed" v="240" />
<tag k="operator" v="東日本旅客鉄道" />
<tag k="frequency" v="50" />
<tag k="highspeed" v="yes" />
<tag k="source_ref" v="http://nlftp.mlit.go.jp/ksj/jpgis/datalist/KsjTmplt-N02-v1_1.html" />
<tag k="start_date" v="1982-11-15" />
<tag k="electrified" v="contact_line" />
<tag k="operator:en" v="East Japan Railway" />
<tag k="operator:ja" v="東日本旅客鉄道" />
</way>
<location lat="37.6485805" lon="138.9391245" />
</error>
<error class="12">
<node id="3864948042" lat="35.092772" lon="136.9159072" version="4" user="kurikimarron">
<tag k="name" v="ファミリーマート" />
<tag k="shop" v="convenience" />
<tag k="brand" v="ファミリーマート" />
<tag k="source" v="bing" />
<tag k="name:en" v="FamilyMart" />
<tag k="name:ja" v="ファミリーマート" />
<tag k="brand:en" v="FamilyMart" />
<tag k="brand:ja" v="ファミリーマート" />
<tag k="brand:wikidata" v="Q1191685" />
<tag k="brand:wikipedia" v="en:FamilyMart" />
</node>
<way id="182170768" version="3" user="kurikimarron">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="unclassified" />
</way>
<location lat="35.092772" lon="136.9159072" />
</error>
<error class="12">
<node id="4127573727" lat="34.7392026" lon="137.7223656" version="1" user="(yo)">
<tag k="name" v="ミートショップさわき" />
<tag k="shop" v="butcher" />
</node>
<way id="186582534" version="6" user="(yo)">
<tag k="highway" v="unclassified" />
</way>
<location lat="34.7392026" lon="137.7223656" />
</error>
<error class="12">
<node id="4134835867" lat="35.3014042" lon="136.801687" version="3" user="TheBlueRobin">
<tag k="name" v="のいり クレストホール本町" />
<tag k="shop" v="funeral_directors" />
<tag k="phone" v="+81 586 24 1597" />
<tag k="website" v="http://www.noiri.co.jp/funeralhall/honmachi/index.html" />
<tag k="operator" v="株式会社のいり" />
<tag k="addr:city" v="愛知県一宮市大志1-6-21" />
<tag k="addr:postcode" v="491-0852" />
</node>
<way id="666686854" version="1" user="TONKATZ">
<tag k="source" v="Bing" />
<tag k="highway" v="unclassified" />
</way>
<location lat="35.3014042" lon="136.801687" />
</error>
<error class="12">
<node id="4246224127" lat="35.1270791" lon="138.9111079" version="3" user="TheBlueRobin">
<tag k="name" v="箱根ベーカリー" />
<tag k="shop" v="bakery" />
<tag k="cuisine" v="bakery" />
<tag k="website" v="http://xn--lckxe4btdc5595f1q2a.com/" />
<tag k="takeaway" v="yes" />
<tag k="addr:city" v="静岡県三島市一番町" />
<tag k="addr:street" v="１６" />
<tag k="addr:postcode" v="411-0036" />
<tag k="opening_hours" v="06:30-20:00" />
<tag k="addr:housenumber" v="−１" />
</node>
<way id="352810944" version="11" user="fuk">
<tag k="name" v="東海道新幹線" />
<tag k="note" v="National-Land Numerical Information (Railway) 2007, MLIT Japan" />
<tag k="gauge" v="1435" />
<tag k="source" v="KSJ2" />
<tag k="name:de" v="Tōkaidō-Shinkansen" />
<tag k="name:en" v="Tokaido Shinkansen" />
<tag k="name:ja" v="東海道新幹線" />
<tag k="name:ru" v="Токайдо-синкансэн" />
<tag k="note:ja" v="国土数値情報（鉄道データ）平成19年　国土交通省" />
<tag k="railway" v="rail" />
<tag k="service" v="siding" />
<tag k="voltage" v="25000" />
<tag k="KSJ2:LIN" v="東海道新幹線" />
<tag k="operator" v="東海旅客鉄道 (JR Central)" />
<tag k="frequency" v="60" />
<tag k="highspeed" v="yes" />
<tag k="wikipedia" v="ja:東海道新幹線" />
<tag k="name:ja_rm" v="JR Tōkaidō Shinkansen" />
<tag k="source_ref" v="http://nlftp.mlit.go.jp/ksj/jpgis/datalist/KsjTmplt-N02-v1_1.html" />
<tag k="electrified" v="contact_line" />
<tag k="operator:en" v="Central Japan Railway Company" />
<tag k="operator:ja" v="東海旅客鉄道" />
<tag k="name:ja_kana" v="とうかいどうしんかんせん" />
<tag k="operator:ja_rm" v="Tōkai Ryokaku Tetsudō" />
</way>
<location lat="35.1270791" lon="138.9111079" />
</error>
<error class="12">
<node id="4402087090" lat="36.1421589" lon="137.2593098" version="1" user="Rufus Chung">
<tag k="shop" v="beverages" />
<tag k="name:zh" v="茶乃芽" />
</node>
<way id="26830733" version="8" user="Carlos_N">
<tag k="name" v="San-no Machi" />
<tag k="oneway" v="yes" />
<tag k="highway" v="residential" />
<tag k="name:es" v="San-no Machi" />
<tag k="surface" v="paved" />
<tag k="source_1" v="Bing" />
<tag k="name:int_name" v="San-no Machi" />
</way>
<location lat="36.1421589" lon="137.2593098" />
</error>
<error class="12">
<node id="5108929291" lat="36.7686001" lon="137.1365834" version="3" user="Jeffrey Friedl">
<tag k="name" v="LAWSON" />
<tag k="shop" v="convenience" />
<tag k="level" v="0" />
</node>
<way id="140692309" version="9" user="xkrautsc">
<tag k="ref" v="415" />
<tag k="foot" v="yes" />
<tag k="name" v="国道415号" />
<tag k="toll" v="no" />
<tag k="lanes" v="2" />
<tag k="oneway" v="no" />
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="bicycle" v="yes" />
<tag k="highway" v="trunk" />
<tag k="name:en" v="National Route 415" />
<tag k="name:ja" v="国道415号" />
<tag k="surface" v="paved" />
</way>
<location lat="36.7686001" lon="137.1365834" />
</error>
<error class="12">
<node id="5276997922" lat="35.4029499" lon="136.7176457" version="1" user="ef200">
<tag k="name" v="岐阜貨物ターミナル" />
<tag k="shop" v="car_parts" />
</node>
<way id="242306266" version="3" user="Jyunhou">
<tag k="gauge" v="1067" />
<tag k="source" v="GSI/KIBAN 2500; NARO" />
<tag k="railway" v="rail" />
<tag k="service" v="yard" />
<tag k="voltage" v="1500" />
<tag k="frequency" v="0" />
<tag k="electrified" v="contact_line" />
<tag k="railway:traffic_mode" v="freight" />
</way>
<location lat="35.4029499" lon="136.7176457" />
</error>
<error class="12">
<node id="5338502963" lat="35.0522814" lon="136.9956199" version="1" user="Phú Hoàng Gia">
<tag k="name" v="ga zengo" />
<tag k="shop" v="ticket" />
</node>
<way id="405116986" version="3" user="MY5DVrAr0e599Japan">
<tag k="name" v="名古屋鉄道名古屋本線" />
<tag k="note" v="National-Land Numerical Information (Railway) 2007, MLIT Japan" />
<tag k="gauge" v="1067" />
<tag k="usage" v="main" />
<tag k="source" v="KSJ2" />
<tag k="name:en" v="Meitetsu Nagoya Main Line" />
<tag k="name:ja" v="名古屋鉄道名古屋本線" />
<tag k="note:ja" v="国土数値情報（鉄道データ）平成19年　国土交通省" />
<tag k="railway" v="rail" />
<tag k="voltage" v="1500" />
<tag k="KSJ2:LIN" v="名古屋本線" />
<tag k="operator" v="名古屋鉄道" />
<tag k="frequency" v="0" />
<tag k="source_ref" v="http://nlftp.mlit.go.jp/ksj/jpgis/datalist/KsjTmplt-N02-v1_1.html" />
<tag k="electrified" v="contact_line" />
<tag k="operator:ja" v="名古屋鉄道" />
<tag k="passenger_lines" v="2" />
</way>
<location lat="35.0522814" lon="136.9956199" />
</error>
<error class="12">
<node id="5585729117" lat="35.0493689" lon="136.9635053" version="3" user="dddnh">
<tag k="name" v="ボーテ" />
<tag k="shop" v="dry_cleaning" />
</node>
<way id="183680440" version="2" user="TKTH">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="unclassified" />
</way>
<location lat="35.0493689" lon="136.9635053" />
</error>
<error class="12">
<node id="6073262786" lat="35.5307533" lon="138.799399" version="1" user="Tina-kutina">
<tag k="shop" v="variety_store" />
</node>
<way id="109405376" version="1" user="leuchtmetall">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="unclassified" />
<tag k="yh:TYPE" v="その他一般道" />
<tag k="yh:WIDTH" v="3.0m〜5.5m" />
<tag k="yh:STRUCTURE" v="地上" />
<tag k="yh:TOTYUMONO" v="供用中" />
<tag k="yh:WIDTH_RANK" v="3" />
</way>
<location lat="35.5307533" lon="138.799399" />
</error>
<error class="12">
<node id="6438987085" lat="37.9549956" lon="139.265467" version="1" user="Sergey Zaharov">
<tag k="shop" v="car" />
<tag k="name:ru" v="дом" />
</node>
<way id="113569747" version="2" user="niityhksn">
<tag k="source" v="YahooJapan/ALPSMAP" />
<tag k="highway" v="tertiary" />
</way>
<location lat="37.9549956" lon="139.265467" />
</error>
```

**Germany_Saarland**:
```
<error class="12">
<node id="5214774681" lat="49.357491" lon="7.1318516" version="2" user="">
<tag k="shop" v="bakery" />
</node>
<way id="46968823" version="5" user="">
<tag k="lit" v="yes" />
<tag k="name" v="Lessingstraße" />
<tag k="oneway" v="yes" />
<tag k="highway" v="residential" />
<tag k="surface" v="asphalt" />
<tag k="maxspeed" v="30" />
<tag k="source:maxspeed" v="DE:zone30" />
</way>
<location lat="49.357491" lon="7.1318516" />
</error>
```
</details>