@frodrigo 2023-04-24
I am OK on the principle. But perf is import here. It is the most expensive analyzer. Please, can you confirm the extra time is reasonable?