@Famlam 2023-04-25
I'll check it in the coming days whether including the buffer (and hence the extra index) makes a big difference in the performance.