@Famlam 2023-04-26
First I searched for the slowest country to run the sql on... 
  -> Running only sql00, sql01, sql08, sql09 and sql32 (others are irrelevant for this PR)
  -> Excluding the first run (building table buildings) which is very slow
Top 3 (out of a selection of 12 regions that I typically use):
1. japan_chubu (110s)
2. slovenia (41s)
3. usa_iowa (25s)

So I used Japan Chubu for all tests, ran on osm110

### SQL00 (not modified in this PR yet)
```
['Bitmap Heap Scan on ways  (cost=456.96..43658.13 rows=687 width=379) (actual time=1126.929..29696.439 rows=1923548 loops=1)']
["  Recheck Cond: (((tags ? 'highway'::text) AND (tags <> ''::hstore)) OR ((tags ? 'railway'::text) AND (tags <> ''::hstore)))"]
['  Rows Removed by Index Recheck: 1097986']
["  Filter: ((NOT (tags ? 'area:highway'::text)) AND (array_length(nodes, 1) <= 100) AND ((NOT (tags ? 'area'::text)) OR ((tags -> 'area'::text) = 'no'::text)) AND (((tags ? 'highway'::text) AND ((tags -> 'highway'::text) <> ALL ('{planned,proposed,construction,rest_area,razed,no,services,elevator}'::text[]))) OR ((tags ? 'railway'::text) AND ((tags -> 'railway'::text) = ANY ('{rail,tram}'::text[])))) AND (st_npoints(linestring) > 1))"]
['  Rows Removed by Filter: 15279']
['  Heap Blocks: exact=34607 lossy=132515']
['  ->  BitmapOr  (cost=456.96..456.96 rows=12771 width=0) (actual time=1109.386..1109.426 rows=0 loops=1)']
['        ->  Bitmap Index Scan on idx_ways_highway  (cost=0.00..228.31 rows=6386 width=0) (actual time=599.165..599.173 rows=1911556 loops=1)']
["              Index Cond: (tags ? 'highway'::text)"]
['        ->  Bitmap Index Scan on idx_ways_tags  (cost=0.00..228.31 rows=6386 width=0) (actual time=510.219..510.219 rows=56225 loops=1)']
["              Index Cond: (tags ? 'railway'::text)"]
['Planning Time: 2.996 ms']
['Execution Time: 29925.152 ms']
```

### SQL01 (indexing)
Not sure how to get proper stats since EXPLAIN ANALYSE doesn't seem to work on INDEX, but runtime is around 36 seconds

## With buffer
### SQL08 with buffer included
Note, there's a bug in sql08 with the craft/shop/office!=no at the moment... should've used `->` instead of `?`. **This was corrected for the below stats but not yet in the PR!**
```
['Bitmap Heap Scan on nodes  (cost=61.94..6287.75 rows=1572 width=104) (actual time=146.688..4137.130 rows=23709 loops=1)']
["  Recheck Cond: (((tags ? 'craft'::text) AND (tags <> ''::hstore)) OR ((tags ? 'shop'::text) AND (tags <> ''::hstore)) OR ((tags ? 'office'::text) AND (tags <> ''::hstore)))"]
['  Rows Removed by Index Recheck: 11609']
["  Filter: (((NOT (tags ? 'street_vendor'::text)) OR ((tags -> 'street_vendor'::text) = 'no'::text)) AND ((NOT (tags ? 'drive_through'::text)) OR ((tags -> 'drive_through'::text) = 'no'::text)) AND ((NOT (tags ? 'level'::text)) OR ((tags -> 'level'::text) = '0'::text)) AND (((tags ? 'craft'::text) AND ((tags -> 'craft'::text) <> 'no'::text)) OR ((tags ? 'shop'::text) AND ((tags -> 'shop'::text) <> 'no'::text)) OR ((tags ? 'office'::text) AND ((tags -> 'office'::text) <> 'no'::text))))"]
['  Rows Removed by Filter: 444']
['  Heap Blocks: exact=16749']
['  ->  BitmapOr  (cost=61.94..61.94 rows=1587 width=0) (actual time=134.423..134.424 rows=0 loops=1)']
['        ->  Bitmap Index Scan on idx_nodes_tags  (cost=0.00..20.25 rows=529 width=0) (actual time=61.974..61.975 rows=5130 loops=1)']
["              Index Cond: (tags ? 'craft'::text)"]
['        ->  Bitmap Index Scan on idx_nodes_tags  (cost=0.00..20.25 rows=529 width=0) (actual time=53.398..53.398 rows=24219 loops=1)']
["              Index Cond: (tags ? 'shop'::text)"]
['        ->  Bitmap Index Scan on idx_nodes_tags  (cost=0.00..20.25 rows=529 width=0) (actual time=19.047..19.047 rows=7930 loops=1)']
["              Index Cond: (tags ? 'office'::text)"]
['Planning Time: 3.389 ms']
['Execution Time: 4142.646 ms']
```

### SQL09 (indexing) 
Not sure how to get proper stats since EXPLAIN ANALYSE doesn't seem to work on INDEX, but runtime is around 0 (zero) seconds

### SQL32 with buffer included 
```
['Unique  (cost=157900.55..157900.55 rows=1 width=80) (actual time=6843.378..6843.393 rows=24 loops=1)']
['  ->  Sort  (cost=157900.55..157900.55 rows=1 width=80) (actual time=6843.377..6843.381 rows=27 loops=1)']
['        Sort Key: commercial.id, highway.id, commercial.geom']
['        Sort Method: quicksort  Memory: 28kB']
['        ->  Nested Loop  (cost=0.41..157900.54 rows=1 width=80) (actual time=110.967..6843.215 rows=27 loops=1)']
['              ->  Seq Scan on commercial  (cost=0.00..2213.09 rows=23709 width=104) (actual time=0.029..58.190 rows=23709 loops=1)']
['              ->  Index Scan using idx_highway_linestring on highway  (cost=0.41..6.56 rows=1 width=72) (actual time=0.284..0.284 rows=0 loops=23709)']
['                    Index Cond: ((commercial.buffergeom && linestring) AND (commercial.buffergeom && linestring))']
["                    Filter: ((level = '0'::text) AND (commercial.layer = layer) AND ((highway IS NULL) OR (highway = ANY ('{living_street,residential,unclassified,tertiary,tertiary_link,secondary,secondary_link,primary,primary_link,trunk,trunk_link,motorway,motorway_link,busway,bus_guideway}'::text[]))) AND _st_intersects(commercial.buffergeom, linestring))"]
['                    Rows Removed by Filter: 3']
['Planning Time: 1.001 ms']
['Execution Time: 6843.897 ms']
```


## Without buffer
### The code I would've used without buffer
```sql
SELECT
    DISTINCT ON(commercial.id)
    commercial.id,
    highway.id,
    ST_AsText(commercial.geom)
FROM
    {0}nodes AS commercial
    JOIN {1}highway AS highway ON
        highway.level = '0' AND
        (
          (commercial.tags?'layer' AND commercial.tags->'layer' = highway.layer) OR
          (NOT commercial.tags?'layer' AND highway.layer = '0')
        ) AND
        commercial.geom && highway.linestring AND
        ST_Intersects(ST_Buffer(commercial.geom::geography, 0.9)::geometry, highway.linestring) AND
        (
            -- exclude shops nearly touching service roads (e.g. unmapped drivethroughs), pedestrian ways (e.g. shopping streets), ...
            highway.highway IS NULL OR -- railways are also in table highway
            highway.highway IN ('living_street', 'residential', 'unclassified', 'tertiary', 'tertiary_link',
                              'secondary', 'secondary_link', 'primary', 'primary_link', 'trunk', 'trunk_link',
                              'motorway', 'motorway_link', 'busway', 'bus_guideway')
        )
WHERE
    commercial.tags != ''::hstore AND
    (
      (commercial.tags?'craft' AND commercial.tags->'craft'!='no') OR
      (commercial.tags?'shop' AND commercial.tags->'shop'!='no') OR
      (commercial.tags?'office' AND commercial.tags->'office'!='no')
    ) AND
    (NOT commercial.tags?'street_vendor' OR commercial.tags->'street_vendor' = 'no') AND
    (NOT commercial.tags?'drive_through' OR commercial.tags->'drive_through' = 'no') AND
    (NOT commercial.tags?'level' OR commercial.tags->'level' = '0')
ORDER BY
    commercial.id,
    highway.id,
    commercial.geom
```
(Note: here I wouldn't use sql08+sql09 anymore since here I only use the buffer once instead of twice; no benefit in pre-calculating it anymore. So this would replace queries sql08, sql09 and sql32)

### The output 
```
['Unique  (cost=19690.74..19690.75 rows=1 width=80) (actual time=17398.491..17398.503 rows=24 loops=1)']
['  ->  Sort  (cost=19690.74..19690.74 rows=1 width=80) (actual time=17398.490..17398.495 rows=27 loops=1)']
['        Sort Key: commercial.id, highway.id, commercial.geom']
['        Sort Method: quicksort  Memory: 28kB']
['        ->  Nested Loop  (cost=62.37..19690.73 rows=1 width=80) (actual time=382.450..17398.378 rows=27 loops=1)']
['              ->  Bitmap Heap Scan on nodes commercial  (cost=61.94..6240.59 rows=1572 width=45) (actual time=107.744..1142.522 rows=23709 loops=1)']
["                    Recheck Cond: (((tags ? 'craft'::text) AND (tags <> ''::hstore)) OR ((tags ? 'shop'::text) AND (tags <> ''::hstore)) OR ((tags ? 'office'::text) AND (tags <> ''::hstore)))"]
['                    Rows Removed by Index Recheck: 11609']
["                    Filter: (((NOT (tags ? 'street_vendor'::text)) OR ((tags -> 'street_vendor'::text) = 'no'::text)) AND ((NOT (tags ? 'drive_through'::text)) OR ((tags -> 'drive_through'::text) = 'no'::text)) AND ((NOT (tags ? 'level'::text)) OR ((tags -> 'level'::text) = '0'::text)) AND (((tags ? 'craft'::text) AND ((tags -> 'craft'::text) <> 'no'::text)) OR ((tags ? 'shop'::text) AND ((tags -> 'shop'::text) <> 'no'::text)) OR ((tags ? 'office'::text) AND ((tags -> 'office'::text) <> 'no'::text))))"]
['                    Rows Removed by Filter: 444']
['                    Heap Blocks: exact=16749']
['                    ->  BitmapOr  (cost=61.94..61.94 rows=1587 width=0) (actual time=98.529..98.531 rows=0 loops=1)']
['                          ->  Bitmap Index Scan on idx_nodes_tags  (cost=0.00..20.25 rows=529 width=0) (actual time=21.107..21.108 rows=5130 loops=1)']
["                                Index Cond: (tags ? 'craft'::text)"]
['                          ->  Bitmap Index Scan on idx_nodes_tags  (cost=0.00..20.25 rows=529 width=0) (actual time=51.215..51.215 rows=24219 loops=1)']
["                                Index Cond: (tags ? 'shop'::text)"]
['                          ->  Bitmap Index Scan on idx_nodes_tags  (cost=0.00..20.25 rows=529 width=0) (actual time=26.202..26.203 rows=7930 loops=1)']
["                                Index Cond: (tags ? 'office'::text)"]
['              ->  Index Scan using idx_highway_linestring on highway  (cost=0.44..8.54 rows=1 width=72) (actual time=0.570..0.570 rows=0 loops=23709)']
["                    Index Cond: ((commercial.geom && linestring) AND ((geography(st_transform(st_buffer(st_transform(geometry((commercial.geom)::geography), _st_bestsrid((commercial.geom)::geography, (commercial.geom)::geography)), '0.9'::double precision), 4326)))::geometry && linestring))"]
["                    Filter: ((level = '0'::text) AND (((commercial.tags ? 'layer'::text) AND ((commercial.tags -> 'layer'::text) = layer)) OR ((NOT (commercial.tags ? 'layer'::text)) AND (layer = '0'::text))) AND ((highway IS NULL) OR (highway = ANY ('{living_street,residential,unclassified,tertiary,tertiary_link,secondary,secondary_link,primary,primary_link,trunk,trunk_link,motorway,motorway_link,busway,bus_guideway}'::text[]))) AND _st_intersects((geography(st_transform(st_buffer(st_transform(geometry((commercial.geom)::geography), _st_bestsrid((commercial.geom)::geography, (commercial.geom)::geography)), '0.9'::double precision), 4326)))::geometry, linestring))"]
['                    Rows Removed by Filter: 3']
['Planning Time: 1.011 ms']
['Execution Time: 17398.911 ms']
```

## Difference with / without buffer
sql00 + sql01 are unaffected hence excluded from the comparison

With buffer (sql08 + sql09 + sql32) 
 - runtime: 4143ms + nihil + 6844ms = 10987ms
 - cost: 6287 + unknown + 157900 = 164187 (not sure if I'm actually allowed to sum these numbers)

Without buffer (sql32new) 
 - runtime: 17399ms
 - cost: 19690
 
I guess runtime is the thing to look at here?

---


## Effect of SQL00 
I also had a little look at sql00. It is largely a copy of table `highways`, but not using that table. If I add table highways to the requirements of this analyser, I can get sql00 to:
```
['Append  (cost=0.00..216515.71 rows=634271 width=475) (actual time=0.388..9440.261 rows=1923548 loops=1)']
['  ->  Seq Scan on highways ways  (cost=0.00..183921.86 rows=634264 width=475) (actual time=0.387..6707.220 rows=1899219 loops=1)']
["        Filter: ((NOT is_construction) AND (NOT is_area) AND (highway <> 'elevator'::text) AND (NOT (tags ? 'area:highway'::text)) AND (array_length(nodes, 1) <= 100))"]
['        Rows Removed by Filter: 11860']
['  ->  Bitmap Heap Scan on ways ways_1  (cost=228.31..23079.79 rows=7 width=379) (actual time=538.904..2534.325 rows=24329 loops=1)']
["        Recheck Cond: ((tags ? 'railway'::text) AND (tags <> ''::hstore))"]
['        Rows Removed by Index Recheck: 28929']
["        Filter: (((tags -> 'railway'::text) = ANY ('{rail,tram}'::text[])) AND (array_length(nodes, 1) <= 100) AND ((NOT (tags ? 'area'::text)) OR ((tags -> 'area'::text) = 'no'::text)) AND (st_npoints(linestring) > 1))"]
['        Rows Removed by Filter: 2967']
['        Heap Blocks: exact=19679']
['        ->  Bitmap Index Scan on idx_ways_tags  (cost=0.00..228.31 rows=6386 width=0) (actual time=529.512..529.513 rows=56225 loops=1)']
["              Index Cond: (tags ? 'railway'::text)"]
['Planning Time: 5.507 ms']
['Execution Time: 9586.665 ms']
```

which is about 3 times faster. (It of course assumes another analyser already build table `highways`, but that table will be build anyway in a full run).
I used the following query as "new" sql00:
```sql
CREATE TEMP TABLE {0}highway AS
SELECT
    id,
    linestring,
    linestring_proj,
    nodes,
    highway,
    coalesce(tags->'level', '0') AS level,
    CASE
        WHEN tags?'layer' THEN tags->'layer'
        WHEN tags->'tunnel' != 'no' THEN 'tunnel'
        WHEN tags->'bridge' != 'no' THEN 'bridge'
        ELSE '0'
    END AS layer,
    tags ?| ARRAY['ford', 'flood_prone'] AS onwater
FROM
    {0}highways AS ways
WHERE
    NOT is_construction AND
    NOT is_area AND
    highway != 'elevator' AND
    NOT tags?'area:highway' AND
    array_length(nodes, 1) <= 100 -- Large ways have too big bbox
UNION ALL
SELECT
    id,
    linestring,
    ST_Transform(linestring, {1}) AS linestring_proj,
    nodes,
    NULL AS highway,
    coalesce(tags->'level', '0') AS level,
    CASE
        WHEN tags?'layer' THEN tags->'layer'
        WHEN tags->'tunnel' != 'no' THEN 'tunnel'
        WHEN tags->'bridge' != 'no' THEN 'bridge'
        ELSE '0'
    END AS layer,
    tags ?| ARRAY['ford', 'flood_prone'] AS onwater
FROM
    {0}ways AS ways
WHERE
    tags != ''::hstore AND
    tags?'railway' AND
    tags->'railway' IN ('rail', 'tram') AND
    (NOT tags?'area' OR tags->'area' = 'no') AND
    array_length(nodes, 1) <= 100 AND -- Large ways have too big bbox
    ST_NPoints(linestring) > 1
```

---
What's your opinion on:
1. using buffer (current implementation) or no buffer (no buffer = slower, but lower cost)
2. using the current sql00 or one based on `highways` UNION'ed with a query on railways (current = slower, but lower cost)